UART Test
Added UART Node and attempted to configure Dockerfiles and mounts. This code outlines a basic demo requiring UART initialization for the TX2 hardware. The UART Node was separated out for testing the ability to communicate without interference from other nodes. This code will be relatively easy to develop further with access to the hardware.